+++ Assets/pallete/tile_asset.meta
@@ -1,10 +1,10 @@
 fileFormatVersion: 2
-guid: 1aa858a5b73d01349a2b9d68184f1a0e
-timeCreated: 1517079202
+guid: 8d31481d8a7b7204b979e258daeaa6b4
+folderAsset: yes
+timeCreated: 1517085517
 licenseType: Free
-NativeFormatImporter:
+DefaultImporter:
   externalObjects: {}
-  mainObjectFileID: 11400000
   userData: 
   assetBundleName: 
   assetBundleVariant:
