+++ Assets/pallete/tile_asset/floor_temp.asset.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: 313dd3f6d426858409f00d52a31c723f
+timeCreated: 1517086307
+licenseType: Free
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
