+++ Assets/pallete/tile_asset/ground_temp.asset
