+++ Assets/pallete/tile_asset/laser_x.asset
@@ -9,9 +9,9 @@ MonoBehaviour:
   m_Enabled: 1
   m_EditorHideFlags: 1
   m_Script: {fileID: 13312, guid: 0000000000000000e000000000000000, type: 0}
-  m_Name: microwave_temp
+  m_Name: laser_x
   m_EditorClassIdentifier: 
-  m_Sprite: {fileID: 21300000, guid: e294e8d7901dcbf4c97d3a31d2230f44, type: 3}
+  m_Sprite: {fileID: 21300000, guid: c6fb09c86b91c0645a024d9a9c6937f3, type: 3}
   m_Color: {r: 1, g: 1, b: 1, a: 1}
   m_Transform:
     e00: 1
