+++ Assets/pallete/tile_asset/laser_x.asset.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: df4e5b4d67a5e0c4cb9338b765182f09
-timeCreated: 1517079095
+guid: e277aaf8c3e49004f80133b9ada6f502
+timeCreated: 1517088390
 licenseType: Free
 NativeFormatImporter:
   externalObjects: {}
