+++ Assets/pallete/tile_asset/laser_y.asset
@@ -9,9 +9,9 @@ MonoBehaviour:
   m_Enabled: 1
   m_EditorHideFlags: 1
   m_Script: {fileID: 13312, guid: 0000000000000000e000000000000000, type: 0}
-  m_Name: table_temp
+  m_Name: laser_y
   m_EditorClassIdentifier: 
-  m_Sprite: {fileID: 21300000, guid: e184840fe78b55546b56d7145bf2cfc5, type: 3}
+  m_Sprite: {fileID: 21300000, guid: ebbda159cde159f40b0f5610eda4b749, type: 3}
   m_Color: {r: 1, g: 1, b: 1, a: 1}
   m_Transform:
     e00: 1
