+++ Assets/pallete/tile_asset/laser_y.asset.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: a1a5ee4c6254e8f44aa73090e1712c1f
-timeCreated: 1517078701
+guid: f1497d96ad335a6408b58213c23470c8
+timeCreated: 1517088437
 licenseType: Free
 NativeFormatImporter:
   externalObjects: {}
