--- Assets/sprite_movement.cs.meta
@@ -1,13 +1,0 @@
-fileFormatVersion: 2
-guid: 25aaa65b75e82ea489e69e7ac9d499fd
-timeCreated: 1517036522
-licenseType: Free
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
